Refactoring: since data types < int are annoying in java, everything (except for booleans) has been changed to int

+++ yane-core/src/main/java/de/turing85/yane/api/CpuBus.java
@@ -1,6 +1,6 @@
 package de.turing85.yane.api;
 
 public interface CpuBus {
-  void write(short address, byte data);
-  byte read(short address);
+  void write(int address, int data);
+  int read(int address);
 }
